Unwatch

+++ flaw.js
@@ -1,8 +1,10 @@
 var Discord = require("discord.js");
 var config = require("./config.json");
 var fs = require("fs");
+var wfa = require("write-file-atomic").sync;
+var watchlist = {}
 var bot = new Discord.Client();
-const prefix = "~~~";
+const prefix = "~";
 
 bot.on("ready", () => {
     console.log("AFlaw is ready");
@@ -10,43 +12,66 @@ bot.on("ready", () => {
     console.log(`Username: ${bot.user.username}`);
 });
 
-bot.on("disconnect", () => {
-    console.log("AFlaw is disconnecting");
+process.on('SIGINT', () => {
+    console.log("AFlaw is exiting");
 });
 
 bot.on("message", msg => {
-    if (msg.author.id != bot.user.id || !msg.content.startsWith(prefix)) return;
+    if (msg.author.id != bot.user.id || msg.content.substring(0, 1) != prefix) return;
 
     let messageContent = msg.content.split(" ");
     let command = messageContent[0].slice(prefix.length);
     let fromBotChannel = msg.channel.id == config.botTestID;
 
     if (command === "watch") {
-        console.log(`Is someone mentioned" ${msg.mentions.users.size!=0}`)
+        console.log("Watch command");
+        console.log(`Is someone mentioned: ${msg.mentions.users.size!=0}`);
         console.log(messageContent[1]);
         let user = msg.mentions.users.size!=0 ? msg.mentions.users.first() : messageContent[1];
-        msg.delete();
+        if (msg.channel.id != config.botTestID) msg.delete();
         watch(msg.mentions.users.size!=0 ? user.username : user);
-        console.log(`Watching ${msg.mentions.users.size!=0 ? user.username : user}`);
-        msg.author.sendMessage(`Watching ${msg.mentions.users.size!=0 ? user.username : user}`);
+        bot.channels.get(config.botTestID).sendMessage(`Watching ${msg.mentions.users.size!=0 ? user.username : user}`)
+        .then(message => {
+            console.log(`Sent watch message: ${message.content}`);
+            message.delete();
+        })
+        .catch(console.error);
+    }
+
+    if (command === "unwatch") {
+        let user = msg.mentions.users.size!=0 ? msg.mentions.users.first() : messageContent[1];
+        if (msg.channel.id != config.botTestID) msg.delete();
+        unwatch(msg.mentions.users.size!=0 ? user.username : user);
+        bot.channels.get(config.botTestID).sendMessage(`No longer watching ${msg.mentions.users.size!=0 ? user.username : user}`)
+        .then(message => {
+            console.log(`Sent unwatch message: ${message.content}`);
+            message.delete();
+        })
+        .catch(console.error);
     }
 
     if (command === "testDM") {
-        // bot.fetchUser(config.flawBotID);
-        // bot.channels.get(config.botTestID).sendMessage("This is a test DM");
-        // msg.author.sendMessage("This is a test DM");
-        console.log(`Sending test message to: ${bot.users.get(config.flawBotID).username}`);
+        bot.channels.get(config.botTestID).sendMessage("This is a test DM");
+        console.log(`Sending test message`);
     }
 });
 
 bot.on("presenceUpdate", (oldUser, newUser) => {
     try {
-        let users = JSON.parse(fs.readFileSync("./watchlist.json"));
-    if (users[oldUser.user.username]) {
-        if ((oldUser.presence.status === "offline") && (newUser.presence.status != "offline")) {
-            bot.channels.get(config.botTestID).sendMessage(`${oldUser.user.username} is ${newUser.presence.status}`);
+        fs.readFile("./watchlist.json", 'utf8', (err, data) => {
+            if (err) throw err;
+            var userObject = JSON.parse(data);
+            if (userObject[oldUser.user.username]) {
+                if ((oldUser.presence.status != "online") && (newUser.presence.status != "offline")) {
+                    bot.channels.get(config.botTestID).sendMessage(`${oldUser.user.username} is ${newUser.presence.status}`)
+                    .then(message => {
+                      console.log(message);
+                      message.delete();  
+                    })
+                    .catch(console.error);
+                }
         }
-    }
+        });
     } catch (error) {
         console.error(error);
     }
@@ -56,10 +81,27 @@ bot.login(config.userToken);
 
 function watch(user) {
     try {
-        let userObject = JSON.parse(fs.readFileSync("./watchlist.json"));
-        userObject[user] = user;
-        console.log(JSON.stringify(user));
-        fs.writeFileSync("./watchlist.json", JSON.stringify(userObject));
+        fs.readFile("./watchlist.json", 'utf8', (err, data) => {
+            if (err) throw err;
+            let userObject = JSON.parse(data);
+            userObject[user] = user;
+            fs.writeFile("./watchlist.json", JSON.stringify(userObject));
+        });
+        return;
+    } catch (error) {
+        console.error(error);
+    }
+}
+
+function unwatch(user) {
+    try {
+        fs.readFile("./watchlist.json", 'utf8', (err, data) => {
+            if (err) throw err;
+            let userObject = JSON.parse(data);
+            delete userObject[user];
+            fs.writeFile("./watchlist.json", JSON.stringify(userObject));
+        });
+        return;
     } catch (error) {
         console.error(error);
     }
